Implemented exception catching in client and server. Client can now request to disconnect from server safely.
Implemented /quit functionality to client and server. If the server recieves a message consisting of /quit, it will disconnect the client from the server and close the socket, and the client will disconnect as well.

Additionally, error checking was added to the server. Now, if a connection is terminated from the server in an unexpected way, the server removes the connection from it's internal list of active connections and closes the socket. Similarly, error checking was implemented in the client. Although the connection still closes, it prevents display of any error messages to the user.