Message functionality now uses pickle module. Implemented color changes. Updated internal documentation.
Message objects now use the python Pickle module to convert lists to bytes that are sent over the socket connection. This allows for modular messages, which will make further development of features much easier.

Users now have a default color associated with their messages (#000000), and are allowed to change it to a valid hexidecimal color code in a message to the server.

Internal documentation of server and client functions were updated. Documentation of message objects was added to client.py for easy reference.

TODO: Move documentation, further extend server messaging functionality.