Added server input handler & thread. Improved error messaging for users.
The server now has an input handler that allows it to communicate with the users connected to the server. This also allows the server to manually shutdown the server, and to generate the configuration files necessary for connection. The server can also, at runtime, specify which port it is running on.

If a user tries to connect to a non-running server, instead of a hard crash, they are met with an error message explaining the error they encountered.

TODO: Improve error messaging throughout the user side of the program.