Added wrapper function for send handler, log off button
Send handler now has a wrapper function that handles the management of the input field before passing the input to the sending logic of the program.

Log off button has been implemented in the client.

Fixed scaling issue with the scrollbar in the client.

Updated README.md with new documentation.